refactor(web): decompose the trace-detail workspace behind a provider seam
Turn the 4,797-line TraceDetailPage into orchestration (fetch + provider +
layout) by concentrating everything derived from a loaded trace behind one
provider seam and extracting the pure derivations.

- Extract eight pure derivations into web/src/utils, each unit-tested:
  timeline rows/phases, log lines, span metrics, engine state machine,
  running-state summary, lineage chain, replay steps, text download.
- Introduce TraceDetailWorkspaceProvider owning derived analyses,
  URL-derived selection, and the expansion set; section components become
  near-zero-prop context consumers (the old 24-prop workspace is now 2).
- TraceDetailPage.tsx: 4,797 -> 384 lines.

Preserves the recorded decisions: selection stays sourced from the span
URL param (no second copy of state), the expansion reducer stays the sole
open-set owner, reveal stays pure derivation. The full-page behavioral
test passes byte-identical; 49 new unit tests cover the derivations.
