+++ web/src/pages/traceDetail/queryError.ts
@@ -1,0 +1,3 @@
+export function queryErrorMessage(error: unknown): string {
+  return error instanceof Error ? error.message : 'Unknown error';
+}
